requirements.txt
fastapi==0.68.1
psycopg2-binary==2.9.1
sqlmodel==0.0.4
uvicorn==0.15.0
PyYAML
easydict
inotify
Twisted
SQLAlchemy~=1.4.10
httpx==0.20.0
pytest==6.2.5
asyncpg
sqlalchemy_utils
pytest_mock